refactor: 신청자/모임장 management 파일 분리
